% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_classifier}
\alias{check_classifier}
\title{Check whether classifier name is valid. (Case-insensitive)
In case only a substring is given, returns the full name
of the classifier.}
\usage{
check_classifier(classifier)
}
\arguments{
\item{classifier}{\code{\link{character}} \cr
Classifier from tsc_classifiers() or short identifier (substring).}
}
\value{
\code{\link{character}}
}
\description{
Check whether classifier name is valid. (Case-insensitive)
In case only a substring is given, returns the full name
of the classifier.
}
