% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_data_from_path}
\alias{delete_data_from_path}
\title{Delete a .arff file.}
\usage{
delete_data_from_path(data_path)
}
\arguments{
\item{data_path}{[`character`] Path to save data to. If null, a
temporary file is created.}
}
\description{
Delete a .arff file.
}

