% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tsc_classifier_names}
\alias{get_tsc_classifier_names}
\title{Return a vector of possible classifier names}
\usage{
get_tsc_classifier_names()
}
\value{
[`character`]
}
\description{
Return a vector of possible classifier names
}
