% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3_LearnerClassifTSClassification.R
\docType{data}
\name{LearnerClassifTSClassification}
\alias{LearnerClassifTSClassification}
\alias{mlr_learners_classif.tsclassification}
\title{Time Series Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
This class interfaces
Time-Series Classification Bake-off classifiers
Calls \code{\link[tsclassification:TSClassifier]{tsclassification::TSClassifier()}} from package \pkg{tsclassification}.
}
\section{Construction}{
\preformatted{LearnerClassifTSClassification$new()
mlr3::mlr_learners$get("classif.tsclassification")
mlr3::lrn("classif.tsclassification")
}
}

\examples{
\dontrun{
  library(mlr3)
  tsk = mlr_tasks$get("iris")
  lrn = LearnerClassifTSClassification$new()
  # Train the classifier
  lrn$train(tsk)
  # Predict again
  lrn$predict(tsk) 
}

}
\references{
Bagnall et al., 2018, The great time-series classification bakeoff
}
\keyword{datasets}
