% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{par_vals_to_string}
\alias{par_vals_to_string}
\title{Flatten a list of parameter values to a single string}
\usage{
par_vals_to_string(par.vals)
}
\arguments{
\item{par.vals}{[`list`] Parameter values.}
}
\value{
[`character`]
}
\description{
Transforms `list(a = "5")` to a string "a = 5".
Multiple elements are concatenated using a ','.
}
