% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSClassifier.R
\name{resample_tsc}
\alias{resample_tsc}
\title{Resample a time-series classifier.
Can be used to replicate the benchmark in the Time-Series Classification Bake-Off by Bagnall et al. 2017
Currently prints out the resampling accuracy.}
\usage{
resample_tsc(data, target = NULL, classifier, par_vals = NULL,
  model_path = NULL, data_path = NULL, cleanup_data = FALSE)
}
\arguments{
\item{data}{\code{\link{character(1)}} | \code{\link{data.frame}} \cr
Either a path to the dataset or a data.frame that should be saved to disk
for modeling. In case a \code{data.frame} is provided, the dataset is saved to disk
via \code{data_to_path}.}

\item{target}{\code{\link{character(1)}} \cr
Name of the target variable.}

\item{classifier}{\code{\link{character(1)}} \cr
Character describing the classifier. See \code{?tsc_classifiers}.}

\item{par_vals}{\code{\link{list}} \cr
(Optional) Hyperparameters for the models. \cr
See \code{?tsc_classifiers} for hyperparameters.}

\item{model_path}{\code{\link{character(1)}} \cr
Path where the resulting model should be saved to.}

\item{data_path}{\code{\link{character(1)}} \cr
Path where train and test data should be saved to. Defaults to a temporary file.}

\item{cleanup_data}{\code{\link{logical(1)}} \cr
Should the data be deleted from disk after training / prediction?}
}
\value{
NULL, Writes a Java instance of TrainAndPredict to \code{model_path}.
data = data.frame(matrix(rnorm(300), nrow = 30))
data$class = factor(sample(letters\link{1:2}, 10, replace = TRUE))
resample_tsc(data, target = "class", classifier = "weka.classifiers.trees.J48", model_path = tempfile())
}
\description{
Resample a time-series classifier.
Can be used to replicate the benchmark in the Time-Series Classification Bake-Off by Bagnall et al. 2017
Currently prints out the resampling accuracy.
}
