% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{predict_tsc}
\alias{predict_tsc}
\title{Predict using a TrainAndPredict Object.}
\usage{
predict_tsc(model_path, newdata)
}
\arguments{
\item{newdata}{[`data.frame` | `character`] \cr
Either a `data.frame` containing
the data, or the file path. In the latter case, nothing happens and the
file path is only checked for consistency.}
}
\value{
A vector of predictions
}
\description{
Predicts newdata using the model obtained
}
