% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSClassifier.R
\name{predict_tsc}
\alias{predict_tsc}
\title{Predict using a TrainAndPredict Object.}
\usage{
predict_tsc(newdata, target = NULL, model_path, data_path = NULL,
  cleanup_data = FALSE)
}
\arguments{
\item{newdata}{[`data.frame` | `character`] \cr
Either a `data.frame` containing
the data, or a file path to data used for prediction.}

\item{target}{[`character(1)`] \cr
Name of the target variable. Can be `NULL`.}

\item{model_path}{[`character(1)`] \cr
Path where the prediction model should be obtained from.}

\item{data_path}{[`character(1)`] \cr
Path where train and test data should be saved to. Defaults to a temporary file.}

\item{cleanup_data}{[`logical(1)`] \cr
Should newdata be deleted from disk after training?}
}
\value{
[`factor`] Vector of predictions.
}
\description{
Predicts `newdata` using the model obtained during training.
}
\examples{
  data = data.frame(matrix(rnorm(300), nrow = 30))
  data$class = factor(sample(letters[1:2], 10, replace = TRUE))
  model_path = tempfile()
  train_tsc(data, target = "class", classifier = "weka.classifiers.trees.J48",
    model_path = model_path)
  predict_tsc(data, target = "class", model_path = model_path)
}
