% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{train_tsc}
\alias{train_tsc}
\title{Train a time-series classifier}
\usage{
train_tsc(classifier, data, par.vals, model.path)
}
\arguments{
\item{classifier}{Character describing the classifier.}

\item{par.vals}{(Optional) Hyperparameters for the models. Currently not used.}

\item{model.path}{Path where the resulting model should be saved.}

\item{filepath}{Path to the dataset.}
}
\value{
a Java instance of TrainAndPredict
}
\description{
Retrieve a Java instance of SpellCorrector, with the training file
specified. Language model is trained before the instance is returned.
The spell corrector is adapted from Peter Norvig's demonstration.
}
