% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{train_tsc}
\alias{train_tsc}
\title{Train a time-series classifier}
\usage{
train_tsc(data, target = NULL, classifier, par_vals = NULL,
  model_path = NULL, data_path = NULL, cleanup_data = FALSE)
}
\arguments{
\item{data}{[`character(1)`|`data.frame`] \cr
Either a path to the dataset or a data.frame that should be saved to disk
for modeling. In case a `data.frame` is provided, the dataset is saved to disk
via `data_to_path`.}

\item{target}{[`character(1)`] \cr
Name of the target variable.}

\item{classifier}{[`character(1)`] \cr
Character describing the classifier. See `?tsc_classifiers`.}

\item{par_vals}{[`list`] \cr
(Optional) Hyperparameters for the models. \cr
See `?tsc_classifiers` for hyperparameters.}

\item{model_path}{[`character(1)`] \cr
Path where the resulting model should be saved to.}

\item{data_path}{[`character(1)`] \cr
Path where train and test data should be saved to. Defaults to a temporary file.}

\item{cleanup_data}{[`logical(1)`] \cr
Should the data be deleted from disk after training / prediction?}
}
\value{
NULL, Writes a Java instance of TrainAndPredict to `model_path`.
}
\description{
Set the "java.options" option to use a higher memory
if required (e.g. `"-Xmx2048m"`).
}
