% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{train_tsc}
\alias{train_tsc}
\title{Train a time-series classifier}
\usage{
train_tsc(data, target = NULL, classifier, par_vals, model_path,
  cleanup_data = FALSE)
}
\arguments{
\item{data}{[`character(1)`|`data.frame`] \cr
Either a path to the dataset or a data.frame that should be saved to disk
for modeling. In case a `data.frame` is provided, the dataset is saved to disk
via `data_to_path`.}

\item{target}{[`character(1)`] \cr
Name of the target variable.}

\item{classifier}{[`character(1)`] \cr
Character describing the classifier.}

\item{par_vals}{[`list`] \cr
(Optional) Hyperparameters for the models. Currently not used.}

\item{model_path}{[`character(1)`] \cr
Path where the resulting model should be saved.}

\item{cleanup_data}{[`logical(1)`] \cr
Should the data be deleted from disk after training?}
}
\value{
NULL, Writes a Java instance of TrainAndPredict to model_path
}
\description{
Set the "java.options" option to use a higher memory
if required (e.g. `"-Xmx2048m"`).
}
