% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_classifiers.R
\name{tsc_classifiers}
\alias{tsc_classifiers}
\title{Return available classifiers}
\usage{
tsc_classifiers()
}
\value{
[`character`] Names of available classifiers.
}
\description{
Run `tsc_classifiers()` to obtain available classifiers.
The following classifiers are available: \cr
}
