% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_classifiers.R
\name{tsc_classifiers}
\alias{tsc_classifiers}
\title{Return available classifiers}
\usage{
tsc_classifiers()
}
\value{
[`character`]
}
\description{
Run `tsc_classifiers()` to obtain available classifiers.
}
\details{
The following classifiers are available: \cr
**Ensemble Classifiers**
\itemize{

\item `timeseriesweka.classifiers.ElasticEnsemble` \cr
   Combination of nearest neighbour (NN) classifiers that use elastic distance measures \cr
   Hyperparameters: *None*

Baselearners of ElasticEnsemble: \cr

`timeseriesweka.classifiers.ensembles.elastic_ensemble.WDTW1NN` \cr
    Elastic Ensemble of Nearest Neighbour Algorithms: \cr
    Weighted Dynamic Time Warping 1 Neares Neighbour \cr
    Hyperparameters: *None*

\item `timeseriesweka.classifiers.FlatCote` \cr
   Collective of Transformation Ensembles (Bagnall et al.,2015) \cr
   Hyperparameters: *None*

}

**Dictionary based Classifiers** \cr
\itemize{
  \item `timeseriesweka.classifiers.BOSS` \cr
  Bag of SFA Symbols
  Hyperparameters:
  \itemize{
    \item `setMaxEnsembleSize`: `integer(1)` range: [1, Inf]
  }
}

**Shapelet based Classifiers**
\itemize{
  \item `timeseriesweka.classifiers.ShapeletTransformClassifier` \cr
  Shapelet transformation that separates the shapelet discovery from the classifier by
  finding the top k shapelets in a single run \cr
  Hyperparameters:
  \itemize{
    \item `setTransformType`: character(1) \cr
      values: "univariate","uni","shapeletd","shapeleti"
    \item `setNumberOfShapelets`: `integer(1)` range: [1, Inf]
  }
}

**Interval based Classifiers**
\itemize{
  \item `timeseriesweka.classifiers.TSF` \cr
  Time Series Forest (Deng et al.,2013) \cr
  Hyperparameters:
  \itemize{
    \item `setNumTrees`: `integer(1)` range: [1, Inf]
  }
}

**Time-Series Classifiers**
\itemize{
  \item timeseriesweka.classifiers.LearnShapelets \cr
    Learned Shapelets (Grabocka et al., 2014) \cr
    Hyperparameters:
  \item timeseriesweka.classifiers.NN_CID \cr

}

**Weka-based Classifiers** \cr
Several WEKA classifiers have been implemented in the Time-Series Classification
Bake-off. \cr
The use of those classifiers is discouraged from within TSClassification,
but nonetheless implemented for completeness.\cr
We advise to use the official implementations from package [RWeka] for greater
flexibility and improved support for setting hyperparameters. \cr
\itemize{
  \item weka.classifiers.functions.Logistic
  \item weka.classifiers.bayes.BayesNet
  \item weka.classifiers.bayes.NaiveBayes
  \item weka.classifiers.functions.Logistic
  \item weka.classifiers.functions.MultilayerPerceptron
  \item weka.classifiers.functions.SMO
  \item weka.classifiers.meta.RotationForest
  \item weka.classifiers.trees.J48
  \item weka.classifiers.trees.RandomForest
}
}
