% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\docType{data}
\name{TSClassifier}
\alias{TSClassifier}
\title{Time Seriess Classifier}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TSClassifier
}
\description{
Allows training and predicting on time-series classification data.
Members:
  - train()
  - predict()
}
\keyword{datasets}

